Removed Connection to Octokit.GraphQL namespace.
It should be in the base namespace for easy access. Also adjusted the default namespace of Octokit.GraphQL.Core and moved the existing files into a `Core` directory, to preserve the namespace/directory mapping.
