+++ Octokit.GraphQL.Core/Connection.cs
@@ -1,17 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Text;
 using System.Threading.Tasks;
 using Octokit.GraphQL.Core.Builders;
 using Octokit.GraphQL.Core.Deserializers;
 using Octokit.GraphQL.Core.Serializers;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
 
-namespace Octokit.GraphQL.Core
+namespace Octokit.GraphQL
 {
     public class Connection
     {
