+++ Octokit.GraphQL.Core/Core/Builders/QueryEntityBuilders.cs
