+++ Octokit.GraphQL.Core/Core/Deserializers/ResponseDeserializer.cs
