+++ Octokit.GraphQL.Core/Core/Disposable.cs
