+++ Octokit.GraphQL.Core/Core/GraphQLQueryException.cs
