+++ Octokit.GraphQL.Core/Core/IQuery.cs
