+++ Octokit.GraphQL.Core/Core/IQueryEntity.cs
