+++ Octokit.GraphQL.Core/Core/IUnion.cs
