+++ Octokit.GraphQL.Core/Core/Introspection/EnumValue.cs
