+++ Octokit.GraphQL.Core/Core/Introspection/SchemaType.cs
