+++ Octokit.GraphQL.Core/Core/Introspection/TypeKind.cs
