+++ Octokit.GraphQL.Core/Core/QueryEntity.cs
