+++ Octokit.GraphQL.Core/Core/QueryEntityExtensions.cs
