+++ Octokit.GraphQL.Core/Core/QueryProvider.cs
