+++ Octokit.GraphQL.Core/Core/Serializers/QuerySerializer.cs
