+++ Octokit.GraphQL.Core/Core/Syntax/ISelectionSet.cs
