+++ Octokit.GraphQL.Core/Core/Syntax/ISyntaxNode.cs
