+++ Octokit.GraphQL.Core/Core/Syntax/InlineFragment.cs
