+++ Octokit.GraphQL.Core/Core/Syntax/OperationDefinition.cs
