+++ Octokit.GraphQL.Core/Core/Syntax/OperationType.cs
