+++ Octokit.GraphQL.Core/Core/Syntax/SelectionSet.cs
