+++ Octokit.GraphQL.Core/Core/Utilities/ExpressionExtensions.cs
