+++ Octokit.GraphQL.Core/Core/Utilities/StringExtensions.cs
